separate PlotlySurfaceFigure tests from SurfaceFigure tests
